Moved code into new file Harmony-Thing.cs
