+++ Source/CombatExtended/CombatExtended.csproj
@@ -262,6 +262,7 @@
     <Compile Include="Harmony\Harmony-Pawn_HealthTracker.cs" />
     <Compile Include="Harmony\Harmony-SmokepopBelt.cs" />
     <Compile Include="Harmony\Harmony-StunHandler.cs" />
+    <Compile Include="Harmony\Harmony-Thing.cs" />
     <Compile Include="Harmony\Harmony-ThingDefGenerator_Meat.cs" />
     <Compile Include="Harmony\Harmony-ThingOwner.cs" />
     <Compile Include="Harmony\Harmony-TooltipUtility.cs" />
