+++ Source/CombatExtended/Harmony/Harmony-Thing.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Harmony;
+using Verse;
+
+namespace CombatExtended.Harmony
+{
+    [HarmonyPatch(typeof(Thing), "SmeltProducts")]
+    public class Harmony_Thing_SmeltProducts
+    {
+        public static void Postfix(Thing __instance, ref IEnumerable<Thing> __result)
+        {
+            var ammoUser = (__instance as ThingWithComps)?.TryGetComp<CompAmmoUser>();
+
+            if (ammoUser != null && (ammoUser.HasMagazine && ammoUser.CurMagCount > 0 && ammoUser.CurrentAmmo != null))
+            {
+                var ammoThing = ThingMaker.MakeThing(ammoUser.CurrentAmmo, null);
+                ammoThing.stackCount = ammoUser.CurMagCount;
+                __result = __result.Add(ammoThing);
+            }
+        }
+    }
+}
