+++ Source/CombatExtended/Harmony/HarmonyThingDef.cs
@@ -72,20 +72,4 @@ namespace CombatExtended.Harmony
             }
         }
     }
-
-    [HarmonyPatch(typeof(Thing), "SmeltProducts")]
-    public class Harmony_Thing_SmeltProducts
-    {
-        public static void Postfix(Thing __instance, ref IEnumerable<Thing> __result)
-        {
-            var ammoUser = (__instance as ThingWithComps)?.TryGetComp<CompAmmoUser>();
-
-            if (ammoUser != null && (ammoUser.HasMagazine && ammoUser.CurMagCount > 0 && ammoUser.CurrentAmmo != null))
-            {
-                var ammoThing = ThingMaker.MakeThing(ammoUser.CurrentAmmo, null);
-                ammoThing.stackCount = ammoUser.CurMagCount;
-                __result = __result.Add(ammoThing);
-            }
-        }
-    }
 }
